requirements.txt
fastapi==0.95.2
strawberry-graphql==0.96.0
uvicorn==0.22.0
sqlalchemy==1.4.41
python-dotenv>=0.19.0
pytest>=6.2.4
httpx==0.24.1
starlette==0.27.0